feat: Simplify commands
