+++ src/cli/new.rs
@@ -4,20 +4,11 @@
 
 use async_trait::async_trait;
 pub use clap::Args;
-use log;
-use std::fs;
-use std::path::PathBuf;
 
 use crate::cli::CliCommand;
-use crate::config::LENRA_CACHE_DIRECTORY;
 use crate::errors::{Error, Result};
-use crate::git::{get_current_commit, GIT_REPO_REGEX};
-use crate::template::{
-    choose_repository, clone_template, list_templates, TemplateData, TEMPLATE_DATA_FILE,
-    TEMPLATE_GIT_DIR,
-};
-#[cfg(test)]
-use mocktopus::macros::mockable;
+use crate::git;
+use crate::{lenra, template};
 
 use super::CommandContext;
 
@@ -39,48 +30,22 @@ impl CliCommand for New {
     async fn run(&self, _context: CommandContext) -> Result<()> {
         log::debug!("topics {:?}", self.topics);
 
-        let template = if self.topics.len() == 1 && GIT_REPO_REGEX.is_match(self.topics[0].as_str())
-        {
-            self.topics[0].clone()
-        } else {
-            let repos = list_templates(&self.topics).await?;
-            if repos.is_empty() {
-                return Err(Error::NoTemplateFound);
-            } else if repos.len() == 1 {
-                repos[0].url.clone()
+        let template =
+            if self.topics.len() == 1 && git::GIT_REPO_REGEX.is_match(self.topics[0].as_str()) {
+                self.topics[0].clone()
             } else {
-                choose_repository(repos).await?.url
-            }
-        };
+                let repos = template::list_templates(&self.topics).await?;
+                if repos.is_empty() {
+                    return Err(Error::NoTemplateFound);
+                } else if repos.len() == 1 {
+                    repos[0].url.clone()
+                } else {
+                    template::choose_repository(repos).await?.url
+                }
+            };
 
-        clone_template(template.clone(), self.path.clone()).await?;
-
-        // create `.template` file to save template repo url and commit
-        let git_dir = self.path.join(".git");
-        let commit = get_current_commit(Some(git_dir.clone())).await?;
-        TemplateData {
-            template,
-            commit: Some(commit.clone()),
-        }
-        .save_to(&self.path.join(TEMPLATE_DATA_FILE))
-        .await
-        .map_err(Error::from)?;
-
-        create_cache_directories(&self.path, &git_dir)?;
-
-        Ok(())
+        lenra::create_new_project(template.as_str(), &self.path).await
     }
-}
-
-#[cfg_attr(test, mockable)]
-fn create_cache_directories(path: &PathBuf, git_dir: &PathBuf) -> Result<()> {
-    log::debug!("create cache directories");
-    // create the `.lenra` cache directory
-    let cache_dir = path.join(LENRA_CACHE_DIRECTORY);
-    fs::create_dir_all(cache_dir.clone()).unwrap();
-    // move the template `.git` directory
-    fs::rename(git_dir, cache_dir.join(TEMPLATE_GIT_DIR))?;
-    Ok(())
 }
 
 #[cfg(test)]
@@ -92,7 +57,7 @@ mod tests {
     use super::*;
     use crate::{
         cli::{self, Command},
-        git::{self, Repository},
+        git::Repository,
         template,
     };
 
@@ -103,11 +68,8 @@ mod tests {
 
     fn mock_all() {
         template::list_templates.mock_safe(|_| unreachable!());
-        template::clone_template.mock_safe(|_, _| unreachable!());
         template::choose_repository.mock_safe(|_| unreachable!());
-        git::get_current_commit.mock_safe(|_| unreachable!());
-        TemplateData::save_to.mock_safe(|_, _| unreachable!());
-        create_cache_directories.mock_safe(|_, _| unreachable!());
+        lenra::create_new_project.mock_safe(|_, _| unreachable!());
     }
 
     #[tokio::test]
@@ -177,34 +139,12 @@ mod tests {
                 }])
             }))
         });
-        let clone_template_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&clone_template_call_counter);
-        template::clone_template.mock_safe(move |url, _| {
-            let mut num = counter.lock().unwrap();
-            *num += 1;
-            assert_eq!(url, NODE_TEMPLATE_HTTP_URL);
-            MockResult::Return(Box::pin(async move { Ok(()) }))
-        });
-        let get_current_commit_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&get_current_commit_call_counter);
-        git::get_current_commit.mock_safe(move |_| {
-            let mut num = counter.lock().unwrap();
-            *num += 1;
-            MockResult::Return(Box::pin(async move { Ok(COMMIT_NUMBER.to_string()) }))
-        });
-        let save_to_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&save_to_call_counter);
-        TemplateData::save_to.mock_safe(move |_, _| {
+        let create_new_project_call_counter = Rc::new(Mutex::new(0));
+        let counter = Rc::clone(&create_new_project_call_counter);
+        lenra::create_new_project.mock_safe(move |_, _| {
             let mut num = counter.lock().unwrap();
             *num += 1;
             MockResult::Return(Box::pin(async move { Ok(()) }))
-        });
-        let create_cache_directories_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&create_cache_directories_call_counter);
-        create_cache_directories.mock_safe(move |_, _| {
-            let mut num = counter.lock().unwrap();
-            *num += 1;
-            MockResult::Return(Ok(()))
         });
         let result = new
             .run(CommandContext {
@@ -212,10 +152,7 @@ mod tests {
             })
             .await;
         assert_eq!(*list_templates_call_counter.lock().unwrap(), 1);
-        assert_eq!(*clone_template_call_counter.lock().unwrap(), 1);
-        assert_eq!(*get_current_commit_call_counter.lock().unwrap(), 1);
-        assert_eq!(*save_to_call_counter.lock().unwrap(), 1);
-        assert_eq!(*create_cache_directories_call_counter.lock().unwrap(), 1);
+        assert_eq!(*create_new_project_call_counter.lock().unwrap(), 1);
         assert!(result.is_ok());
         Ok(())
     }
@@ -264,34 +201,12 @@ mod tests {
             assert_eq!(repos.len(), 2);
             MockResult::Return(Box::pin(async move { Ok(repos[0].clone()) }))
         });
-        let clone_template_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&clone_template_call_counter);
-        template::clone_template.mock_safe(move |url, _| {
-            let mut num = counter.lock().unwrap();
-            *num += 1;
-            assert_eq!(url, NODE_TEMPLATE_HTTP_URL);
-            MockResult::Return(Box::pin(async move { Ok(()) }))
-        });
-        let get_current_commit_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&get_current_commit_call_counter);
-        git::get_current_commit.mock_safe(move |_| {
-            let mut num = counter.lock().unwrap();
-            *num += 1;
-            MockResult::Return(Box::pin(async move { Ok(COMMIT_NUMBER.to_string()) }))
-        });
-        let save_to_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&save_to_call_counter);
-        TemplateData::save_to.mock_safe(move |_, _| {
+        let create_new_project_call_counter = Rc::new(Mutex::new(0));
+        let counter = Rc::clone(&create_new_project_call_counter);
+        lenra::create_new_project.mock_safe(move |_, _| {
             let mut num = counter.lock().unwrap();
             *num += 1;
             MockResult::Return(Box::pin(async move { Ok(()) }))
-        });
-        let create_cache_directories_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&create_cache_directories_call_counter);
-        create_cache_directories.mock_safe(move |_, _| {
-            let mut num = counter.lock().unwrap();
-            *num += 1;
-            MockResult::Return(Ok(()))
         });
         let result = new
             .run(CommandContext {
@@ -299,10 +214,7 @@ mod tests {
             })
             .await;
         assert_eq!(*list_templates_call_counter.lock().unwrap(), 1);
-        assert_eq!(*clone_template_call_counter.lock().unwrap(), 1);
-        assert_eq!(*get_current_commit_call_counter.lock().unwrap(), 1);
-        assert_eq!(*save_to_call_counter.lock().unwrap(), 1);
-        assert_eq!(*create_cache_directories_call_counter.lock().unwrap(), 1);
+        assert_eq!(*create_new_project_call_counter.lock().unwrap(), 1);
         assert!(result.is_ok());
         Ok(())
     }
@@ -320,44 +232,19 @@ mod tests {
         assert_eq!(new.path, std::path::PathBuf::from("."));
         assert_eq!(new.topics, expected_topics);
         mock_all();
-        let clone_template_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&clone_template_call_counter);
-        template::clone_template.mock_safe(move |url, _| {
-            let mut num = counter.lock().unwrap();
-            *num += 1;
-            assert_eq!(url, NODE_TEMPLATE_HTTP_URL);
-            MockResult::Return(Box::pin(async move { Ok(()) }))
-        });
-        let get_current_commit_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&get_current_commit_call_counter);
-        git::get_current_commit.mock_safe(move |_| {
-            let mut num = counter.lock().unwrap();
-            *num += 1;
-            MockResult::Return(Box::pin(async move { Ok(COMMIT_NUMBER.to_string()) }))
-        });
-        let save_to_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&save_to_call_counter);
-        TemplateData::save_to.mock_safe(move |_, _| {
+        let create_new_project_call_counter = Rc::new(Mutex::new(0));
+        let counter = Rc::clone(&create_new_project_call_counter);
+        lenra::create_new_project.mock_safe(move |_, _| {
             let mut num = counter.lock().unwrap();
             *num += 1;
             MockResult::Return(Box::pin(async move { Ok(()) }))
-        });
-        let create_cache_directories_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&create_cache_directories_call_counter);
-        create_cache_directories.mock_safe(move |_, _| {
-            let mut num = counter.lock().unwrap();
-            *num += 1;
-            MockResult::Return(Ok(()))
         });
         let result = new
             .run(CommandContext {
                 ..Default::default()
             })
             .await;
-        assert_eq!(*clone_template_call_counter.lock().unwrap(), 1);
-        assert_eq!(*get_current_commit_call_counter.lock().unwrap(), 1);
-        assert_eq!(*save_to_call_counter.lock().unwrap(), 1);
-        assert_eq!(*create_cache_directories_call_counter.lock().unwrap(), 1);
+        assert_eq!(*create_new_project_call_counter.lock().unwrap(), 1);
         assert!(result.is_ok());
         Ok(())
     }
@@ -375,44 +262,19 @@ mod tests {
         assert_eq!(new.path, std::path::PathBuf::from("."));
         assert_eq!(new.topics, expected_topics);
         mock_all();
-        let clone_template_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&clone_template_call_counter);
-        template::clone_template.mock_safe(move |url, _| {
-            let mut num = counter.lock().unwrap();
-            *num += 1;
-            assert_eq!(url, NODE_TEMPLATE_SSH_URL);
-            MockResult::Return(Box::pin(async move { Ok(()) }))
-        });
-        let get_current_commit_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&get_current_commit_call_counter);
-        git::get_current_commit.mock_safe(move |_| {
-            let mut num = counter.lock().unwrap();
-            *num += 1;
-            MockResult::Return(Box::pin(async move { Ok(COMMIT_NUMBER.to_string()) }))
-        });
-        let save_to_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&save_to_call_counter);
-        TemplateData::save_to.mock_safe(move |_, _| {
+        let create_new_project_call_counter = Rc::new(Mutex::new(0));
+        let counter = Rc::clone(&create_new_project_call_counter);
+        lenra::create_new_project.mock_safe(move |_, _| {
             let mut num = counter.lock().unwrap();
             *num += 1;
             MockResult::Return(Box::pin(async move { Ok(()) }))
-        });
-        let create_cache_directories_call_counter = Rc::new(Mutex::new(0));
-        let counter = Rc::clone(&create_cache_directories_call_counter);
-        create_cache_directories.mock_safe(move |_, _| {
-            let mut num = counter.lock().unwrap();
-            *num += 1;
-            MockResult::Return(Ok(()))
         });
         let result = new
             .run(CommandContext {
                 ..Default::default()
             })
             .await;
-        assert_eq!(*clone_template_call_counter.lock().unwrap(), 1);
-        assert_eq!(*get_current_commit_call_counter.lock().unwrap(), 1);
-        assert_eq!(*save_to_call_counter.lock().unwrap(), 1);
-        assert_eq!(*create_cache_directories_call_counter.lock().unwrap(), 1);
+        assert_eq!(*create_new_project_call_counter.lock().unwrap(), 1);
         assert!(result.is_ok());
         Ok(())
     }
