+++ src/cli/stop.rs
@@ -1,12 +1,9 @@
-use std::process::Stdio;
-
 use async_trait::async_trait;
 pub use clap::Args;
-use log::warn;
 
 use crate::cli::CliCommand;
-use crate::docker_compose::create_compose_command;
-use crate::errors::{CommandError, Error, Result};
+use crate::errors::Result;
+use crate::lenra;
 
 use super::CommandContext;
 
@@ -16,22 +13,6 @@ pub struct Stop;
 #[async_trait]
 impl CliCommand for Stop {
     async fn run(&self, _context: CommandContext) -> Result<()> {
-        log::info!("Stoping the app");
-
-        let mut command = create_compose_command();
-
-        command
-            .arg("down")
-            .arg("--volumes")
-            .stdout(Stdio::inherit())
-            .stderr(Stdio::inherit());
-
-        log::debug!("cmd: {:?}", command);
-        let output = command.spawn()?.wait_with_output().await?;
-        if !output.status.success() {
-            warn!("An error occured while stoping the docker-compose app");
-            return Err(Error::Command(CommandError { command, output }));
-        }
-        Ok(())
+        lenra::stop_env().await
     }
 }
